Make hash values available in build context
This tiny change makes Fabricator more powerful. For example if I create material for a button element like this:
```
<button class="btn{{#if class}} {{class}}{{/if}}">Button</button>
```
I can then use the button helper like this `{{> button}}` to get a standard button:
```
<button class="btn">Button</button>
```
Or like this `{{> button class="btn-sm"}}` to add a class `btn-sm` and I will get this:
```
<button class="btn">Button</button>
```
This can be useful for adding more than just classes to helpers. The posibilities depend on how complex you want to make materials. But there are lots of sumple use cases.

This paragraph helper will accept an optional content attribute:
```
{{#if content}}
  <p>{{content}}</p>
{{/if}}
{{#unless content}}
  <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam a mi massa....</p>
{{/unless}}
```
Usage:  `{{> paragraph content="This is a short sentence."}}` will result in:
```
<p>This is a short sentence.</p>
```